+++ functions_framework_builder/lib/src/supported_function_type.dart
@@ -68,7 +68,7 @@ class SupportedFunctionType {
     FunctionElement element,
   ) {
     if (element.library.typeSystem.isSubtypeOf(element.type, _type)) {
-      return _TrivialFactoryData(
+      return TrivialFactoryData(
         escapeDartString(targetName),
         '$_constructor('
         '$functionsLibraryPrefix.${element.name},)',
@@ -84,64 +84,11 @@ abstract class FactoryData {
   String get expression;
 }
 
-class _TrivialFactoryData implements FactoryData {
+class TrivialFactoryData implements FactoryData {
   @override
   final String name;
   @override
   final String expression;
 
-  _TrivialFactoryData(this.name, this.expression);
-}
-
-class SupportedMiddlewareType {
-  final String libraryUri;
-  final String typedefName;
-  final String typeDescription;
-
-  final FunctionType _type;
-
-  SupportedMiddlewareType._(
-    this.libraryUri,
-    this.typedefName,
-    this._type,
-  ) : typeDescription = _type.getDisplayString(withNullability: false);
-
-  static Future<SupportedMiddlewareType> create(
-    Resolver resolver,
-    String libraryUri,
-    String typeDefName,
-    String constructor,
-  ) async {
-    final lib = await resolver.libraryFor(
-      AssetId.resolve(Uri.parse(libraryUri)),
-    );
-
-    final handlerTypeAlias =
-        lib.exportNamespace.get(typeDefName) as TypeAliasElement;
-
-    final functionType = handlerTypeAlias.instantiate(
-      typeArguments: [],
-      nullabilitySuffix: NullabilitySuffix.none,
-    );
-
-    return SupportedMiddlewareType._(
-      libraryUri,
-      typeDefName,
-      functionType as FunctionType,
-    );
-  }
-
-  FactoryData? createReference(
-    LibraryElement library,
-    String targetName,
-    FunctionElement element,
-  ) {
-    if (element.library.typeSystem.isSubtypeOf(element.type, _type)) {
-      return _TrivialFactoryData(
-        escapeDartString(targetName),
-        '$functionsLibraryPrefix.${element.name}',
-      );
-    }
-    return null;
-  }
+  TrivialFactoryData(this.name, this.expression);
 }
